+++ lib/td_dd/grants/requests.ex
@@ -391,20 +391,6 @@ defmodule TdDd.Grants.Requests do
     |> enrich()
   end
 
-  defp maybe_update_pending_removal(
-         :grant_removal,
-         %Grant{} = grant,
-         %{status: "approved"},
-         claims
-       ) do
-    {:ok, %{grant: grant}} = Grants.update_grant(grant, %{pending_removal: true}, claims)
-    {:ok, grant}
-  end
-
-  defp maybe_update_pending_removal(_request_type, _grant, _status, _claims) do
-    {:ok, nil}
-  end
-
   def bulk_create_approvals(
         %{user_id: user_id} = claims,
         grant_requests,
@@ -445,9 +431,17 @@ defmodule TdDd.Grants.Requests do
 
     Multi.new()
     |> Multi.insert_all(:approvals, GrantRequestApproval, grant_request_entries,
-      returning: [:id, :grant_request_id, :comment, :user_id, :role, :is_rejection]
+      returning: [
+        :id,
+        :grant_request_id,
+        :comment,
+        :user_id,
+        :role,
+        :is_rejection
+      ]
     )
-    |> bulk_maybe_insert_status(grant_request_entries)
+    |> bulk_maybe_insert_status
+    |> bulk_maybe_update_pending_removal(grant_requests)
     |> Multi.run(:audit, Audit, :grant_request_bulk_approval_created, [])
     |> Repo.transaction()
     |> on_upsert()
@@ -468,26 +462,89 @@ defmodule TdDd.Grants.Requests do
   def reindex_on_data_structure_update(data_structure_ids),
     do: reindex_on_data_structure_update([data_structure_ids])
 
-  defp bulk_maybe_insert_status(multi, grant_requests_entries) do
-    Multi.run(multi, :statuses, fn _, _ ->
-      status_entries =
-        grant_requests_entries
+  defp bulk_maybe_insert_status(multi) do
+    Multi.run(multi, :statuses, fn _, changes ->
+      approval_entries =
+        changes
+        |> Map.get(:approvals)
+        |> elem(1)
         |> Enum.reduce([], fn grant_request, acc ->
           [validate_grant_request_status(grant_request) | acc]
         end)
         |> Enum.reject(&is_nil/1)
 
-      if Enum.any?(status_entries, &(&1 == :error)) do
+      if Enum.any?(approval_entries, &(&1 == :error)) do
         {:error, :insert_status}
       else
         result =
-          Repo.insert_all(GrantRequestStatus, status_entries,
+          Repo.insert_all(GrantRequestStatus, approval_entries,
             returning: [:id, :status, :grant_request_id]
           )
 
         {:ok, result}
       end
     end)
+  end
+
+  defp bulk_maybe_update_pending_removal(
+         multi,
+         grant_requests
+       ) do
+    Multi.run(multi, :update_pending_removal_grants, fn _, %{statuses: {_, statuses}} ->
+      grant_ids =
+        statuses
+        |> Enum.filter(fn %{status: status} -> status == "approved" end)
+        |> Enum.reduce([], fn
+          %{grant_request_id: grant_request_id}, acc ->
+            get_grant_from_grant_request(grant_requests, grant_request_id, acc)
+        end)
+
+      {count, _} =
+        Grant
+        |> where([g], g.id in ^grant_ids)
+        |> Repo.update_all(set: [pending_removal: true])
+
+      {:ok, {count, grant_ids}}
+    end)
+  end
+
+  defp get_grant_from_grant_request(grant_requests, grant_request_id, acc) do
+    grant_requests
+    |> Enum.find(fn gr ->
+      gr.id == grant_request_id and to_string(gr.request_type) == "grant_removal"
+    end)
+    |> case do
+      nil ->
+        acc
+
+      %{grant_id: grant_id} ->
+        [grant_id | acc]
+    end
+  end
+
+  defp maybe_update_pending_removal(
+         "grant_removal",
+         %Grant{} = grant,
+         status,
+         claims
+       ) do
+    maybe_update_pending_removal(:grant_removal, grant, status, claims)
+  end
+
+  defp maybe_update_pending_removal(
+         :grant_removal,
+         %Grant{} = grant,
+         %{status: status},
+         claims
+       )
+       when status in ["approved"] do
+    {:ok, %{grant: grant}} = Grants.update_grant(grant, %{pending_removal: true}, claims)
+
+    {:ok, grant}
+  end
+
+  defp maybe_update_pending_removal(_request_type, _grant, _status, _claims) do
+    {:ok, nil}
   end
 
   defp validate_grant_request_status(%{is_rejection: true} = grant_request) do