+++ test/td_dd/data_structures/search_test.exs
@@ -13,11 +13,6 @@ defmodule TdDd.DataStructures.SearchTest do
 
   setup :verify_on_exit!
 
-  setup do
-    start_supervised!(TdCore.Search.Cluster)
-    :ok
-  end
-
   describe "get_filter_values/3" do
     for role <- ["admin", "service"] do
       @tag authentication: [role: role]