+++ test/td_dd/grants/requests_test.exs
@@ -4,7 +4,7 @@ defmodule TdDd.Grants.RequestsTest do
 
   import TdDd.TestOperators
 
-  alias TdCore.Search.MockIndexWorker
+  alias TdCore.Search.IndexWorkerMock
   alias TdDd.Grants
   alias TdDd.Grants.GrantRequest
   alias TdDd.Grants.GrantRequestApproval
@@ -16,9 +16,6 @@ defmodule TdDd.Grants.RequestsTest do
   @valid_metadata %{"list" => "one", "string" => "bar"}
 
   setup tags do
-    start_supervised!(TdCore.Search.Cluster)
-    start_supervised!(TdCore.Search.IndexWorker)
-
     case Map.get(tags, :role, "user") do
       "admin" ->
         claims = build(:claims, role: "admin")
@@ -47,6 +44,7 @@ defmodule TdDd.Grants.RequestsTest do
     end
 
     test "create_grant_request_group/2 with valid data creates a grant_request_group" do
+      IndexWorkerMock.clear()
       %{id: domain_id} = CacheHelpers.insert_domain()
       %{id: data_structure_id} = insert(:data_structure, domain_ids: [domain_id])
       %{user_id: user_id} = build(:claims)
@@ -70,7 +68,7 @@ defmodule TdDd.Grants.RequestsTest do
 
       assert %{status: "pending"} = Repo.get_by!(GrantRequestStatus, grant_request_id: request_id)
 
-      assert [{:reindex, :grant_requests, [^request_id]}] = MockIndexWorker.calls()
+      assert [{:reindex, :grant_requests, [^request_id]}] = IndexWorkerMock.calls()
     end
 
     test "create_grant_request_group/2 with modification_grant" do
@@ -371,13 +369,14 @@ defmodule TdDd.Grants.RequestsTest do
     end
 
     test "delete_grant_request/1 deletes the grant_request", %{claims: claims} do
+      IndexWorkerMock.clear()
       %{id: grant_request_id} = grant_request = insert(:grant_request)
       assert {:ok, %GrantRequest{}} = Requests.delete_grant_request(grant_request)
 
       assert_raise Ecto.NoResultsError, fn ->
         Requests.get_grant_request!(grant_request.id, claims)
 
-        assert MockIndexWorker.calls() == [{:delete_grant_request, [grant_request_id]}]
+        assert IndexWorkerMock.calls() == [{:delete_grant_request, [grant_request_id]}]
       end
     end
 
@@ -415,6 +414,8 @@ defmodule TdDd.Grants.RequestsTest do
       domain_id: domain_id,
       request: request
     } do
+      IndexWorkerMock.clear()
+
       CacheHelpers.put_grant_request_approvers([
         %{user_id: user_id, domain_id: domain_id, role: "approver"}
       ])
@@ -425,7 +426,7 @@ defmodule TdDd.Grants.RequestsTest do
       assert %{is_rejection: false, user: user} = approval
       assert %{id: ^user_id, user_name: _} = user
 
-      assert [{:reindex, :grant_requests, [_]}] = MockIndexWorker.calls()
+      assert [{:reindex, :grant_requests, [_]}] = IndexWorkerMock.calls()
     end
 
     test "admin can approve a grant request without having the role", %{
@@ -490,6 +491,8 @@ defmodule TdDd.Grants.RequestsTest do
       grant: %{id: grant_id} = grant,
       request: request
     } do
+      IndexWorkerMock.clear()
+
       CacheHelpers.put_grant_request_approvers([
         %{user_id: user_id, domain_id: domain_id, role: "approver"}
       ])
@@ -512,11 +515,12 @@ defmodule TdDd.Grants.RequestsTest do
       assert %{pending_removal: true} = Grants.get_grant(grant_id)
 
       assert [{:reindex, :grants, [_]}, {:reindex, :grant_requests, [_]}] =
-               MockIndexWorker.calls()
+               IndexWorkerMock.calls()
     end
   end
 
   defp setup_request_grant_removal(%{claims: %{user_id: user_id}}) do
+    IndexWorkerMock.clear()
     %{id: domain_id} = CacheHelpers.insert_domain()
     CacheHelpers.insert_user(user_id: user_id)
 
@@ -540,6 +544,7 @@ defmodule TdDd.Grants.RequestsTest do
   end
 
   defp setup_request_access(%{claims: %{user_id: user_id}}) do
+    IndexWorkerMock.clear()
     %{id: domain_id} = CacheHelpers.insert_domain()
     CacheHelpers.insert_user(user_id: user_id)