+++ test/td_dd/grants/statuses_test.exs
@@ -5,26 +5,28 @@ defmodule TdDd.Grants.StatusesTest do
   alias TdDd.Grants.GrantRequestStatus
   alias TdDd.Grants.Statuses
 
-  alias TdCore.Search.MockIndexWorker
+  alias TdCore.Search.IndexWorkerMock
 
   setup do
-    start_supervised!(TdCore.Search.Cluster)
-    start_supervised!(TdCore.Search.IndexWorker)
     [grant_request: insert(:grant_request, current_status: "approved")]
   end
 
   describe "Requests.create_grant_request_status/2" do
     test "fails with non valid current grant request status", %{grant_request: request} do
+      IndexWorkerMock.clear()
+
       assert {:error, :grant_request_status,
               %Changeset{errors: [status: {"invalid status change", _}]},
               _} = Statuses.create_grant_request_status(request, "not_valid_status", 0)
 
-      assert MockIndexWorker.calls() == []
+      assert IndexWorkerMock.calls() == []
     end
 
     test "creates grant request status with valid status change", %{
       grant_request: %{id: id} = request
     } do
+      IndexWorkerMock.clear()
+
       assert {:ok,
               %{
                 grant_request_status: %GrantRequestStatus{
@@ -33,10 +35,11 @@ defmodule TdDd.Grants.StatusesTest do
                 }
               }} = Statuses.create_grant_request_status(request, "processing", 0)
 
-      assert [{:reindex, :grant_requests, [^id]}] = MockIndexWorker.calls()
+      assert [{:reindex, :grant_requests, [^id]}] = IndexWorkerMock.calls()
     end
 
     test "allows to create failed status with a reason if current is processing" do
+      IndexWorkerMock.clear()
       %{id: id} = request = insert(:grant_request, current_status: "processing")
       reason = "failed reason"
 
@@ -49,7 +52,7 @@ defmodule TdDd.Grants.StatusesTest do
                 }
               }} = Statuses.create_grant_request_status(request, "failed", 0, reason)
 
-      assert [{:reindex, :grant_requests, [^id]}] = MockIndexWorker.calls()
+      assert [{:reindex, :grant_requests, [^id]}] = IndexWorkerMock.calls()
     end
   end
 end