+++ test/td_dd_web/controllers/data_structure_filter_controller_test.exs
@@ -10,13 +10,6 @@ defmodule TdDdWeb.DataStructureFilterControllerTest do
     }
   }
 
-  setup_all do
-    start_supervised!(TdCore.Search.Cluster)
-    start_supervised!(TdCore.Search.IndexWorker)
-
-    :ok
-  end
-
   setup :verify_on_exit!
 
   describe "index" do