+++ test/td_dq/rules/rule_results/bulk_load_test.exs
@@ -10,8 +10,6 @@ defmodule TdDq.RuleResults.BulkLoadTest do
   @stream TdCache.Audit.stream()
 
   setup_all do
-    start_supervised!(TdCore.Search.Cluster)
-    start_supervised!(TdCore.Search.IndexWorker)
     start_supervised(TdDq.MockRelationCache)
     start_supervised(TdDq.Cache.RuleLoader)