+++ test/td_dq_web/controllers/implementation_result_controller_test.exs
@@ -2,14 +2,7 @@ defmodule TdDqWeb.ImplementationResultControllerTest do
   use TdDqWeb.ConnCase
   use PhoenixSwagger.SchemaTest, "priv/static/swagger_dq.json"
 
-  alias TdCore.Search.MockIndexWorker
-
-  setup_all do
-    start_supervised!(TdCore.Search.Cluster)
-    start_supervised!(TdCore.Search.IndexWorker)
-
-    :ok
-  end
+  alias TdCore.Search.IndexWorkerMock
 
   setup context do
     %{id: domain_id} =
@@ -167,7 +160,7 @@ defmodule TdDqWeb.ImplementationResultControllerTest do
 
     @tag authentication: [role: "service"]
     test "reindexes rule and implementation after creation", %{conn: conn} do
-      MockIndexWorker.clear()
+      IndexWorkerMock.clear()
 
       %{id: rule_id} = rule = insert(:rule)
 
@@ -187,7 +180,7 @@ defmodule TdDqWeb.ImplementationResultControllerTest do
       assert [
                {:reindex, :rules, [^rule_id]},
                {:reindex, :implementations, [^implementation_id]}
-             ] = MockIndexWorker.calls()
+             ] = IndexWorkerMock.calls()
     end
 
     @tag authentication: [role: "admin"]