+++ test/td_dq_web/controllers/rule_filter_controller_test.exs
@@ -8,11 +8,6 @@ defmodule TdDqWeb.RuleFilterControllerTest do
 
   setup :verify_on_exit!
 
-  setup do
-    start_supervised!(TdCore.Search.Cluster)
-    :ok
-  end
-
   describe "index" do
     @tag authentication: [role: "admin"]
     test "maps filters from request parameters", %{conn: conn} do